src/foo/requirements.txt
Flask==2.0.1
grpcio==1.38.0
grpcio-tools==1.38.0
protobuf==3.17.1
python-json-logger==2.0.1
json-logging==1.3.0


# flask_caching==1.10.1